Add background health research task.
This adds a background health research task with a delay start of 30 seconds. On version of iOS older than 17.0 it will fall back to BGAppRefreshTask. In addition, a longer/heavier (min 12 hour delay task)
is scheduled to run in the background as an additional supplement to fetching data from SensorKit.